Adding remaining changes, indentation, removing hardcoded path for Sqlite database

+++ database/migrations/2021_12_21_054432_create_coffee-bean_table.php
@@ -6,18 +6,14 @@ use Illuminate\Support\Facades\Schema;
 
 class CreateCoffeeBeanTable extends Migration
 {
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
+
     public function up()
     {
         Schema::create('coffeeBeanTable', function (Blueprint $table) {
             $table->id('id');
             $table->string('Name', 30);
             $table->enum('Caffeine_Level', ['decaf', 'low', 'medium', 'high']);
-            $table->integer('Cost_lb')->default(0);
+            $table->double('Cost_lb')->default(0.0);
             $table->string('Bean_Type', ['Arabica', 'Robusta', 'Liberca', 'Excelsa', 'Other']);
             $table->enum('Roast', ['dark', 'medium', 'light', 'french']);
             $table->enum('Grind', ['whole-bean', 'fine', 'medium', 'coarse']);
@@ -26,11 +22,6 @@ class CreateCoffeeBeanTable extends Migration
         });
     }
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
     public function down()
     {
         Schema::dropIfExists('coffee_bean');
